+++ nncf/experimental/torch/search_building_blocks/search_graph.py
@@ -264,7 +264,7 @@ def get_merged_original_graph_with_pattern(orig_graph: nx.DiGraph, hw_fused_ops:
         return merged_graph
     # pylint: disable=protected-access
     pattern_fusing_graph =\
-        PatternsManager().get_full_pattern_graph(BackendType.TORCH, TargetDevice.ANY)
+        PatternsManager.get_full_pattern_graph(BackendType.TORCH, TargetDevice.ANY)
     matches = find_subgraphs_matching_pattern(orig_graph, pattern_fusing_graph)
     nx.set_node_attributes(merged_graph, False, SearchGraph.IS_DUMMY_NODE_ATTR)
     nx.set_node_attributes(merged_graph, False, SearchGraph.IS_MERGED_NODE_ATTR)
